refactor: CreditManager class 분리
- main 파일에서 CreditManager class를 분리
- CreditManager class의 프로퍼티(students)와 메서드(except startProgram)를 private으로 변경
